Swatch: Separated Color Space Functionality
I divided specific color space methods into their own files to stay organized.

I was tempted to create separate classes for each color space. I opted not to do this this because it would interfere with the defaults of initialization. That would require the user to specify which color space is being initialized beforehand every time a swatch is created.